+++ examples/data-sources/hcloud_storage_box_subaccount/data-source.tf
@@ -5,6 +5,11 @@ data "hcloud_storage_box_subaccount" "by_id" {
   id             = 2
 }
 
+data "hcloud_storage_box_subaccount" "by_name" {
+  storage_box_id = var.storage_box_id
+  name           = "badger"
+}
+
 data "hcloud_storage_box_subaccount" "by_username" {
   storage_box_id = var.storage_box_id
   username       = "u507137-sub1"